drivers: can: replace canbus_np_context with socket_can_context
Move struct socket_can_context to socket_can_context.h to ensure
struct socket_can_context can be included without replying on
socket_can_generic.h.

Change rx thread of canbus to be preemptive to avoid can frame loss.

Update naming of _ENABLE to _ENABLED to follow convention in
Kconfig.native_posix.
